[fix]: [PL-31077]: New User AccountLevel data and related CRUD API changes (#45273)
* [fix]: [PL-31077]: New User AccountLevel data and related CRUD API changes of User


* [fix]: [PL-31077]: Introduce FF PL_USER_ACCOUNT_LEVEL_DATA_FLOW


* [fix]: [PL-31077]: Remove files added via Audit streaming testing by mistake


* [feat]: [PL-31077]: Merge develop and added new Preference class


* [feat]: [PL-31077]: Reformt


* [feat]: [PL-31077]: Handle deletion


* [feat]: [PL-31077]: Handle deletion


* [feat]: [PL-31077]: Add Copyright statement


* [feat]: [PL-31077]: format check


* [feat]: [PL-31077]: Remove unnecessary comments and fix NPE in Sign up.


* [feat]: [PL-31077]: Remove unnecessary comments and fix NPE in Sign up.


* [feat]: [PL-31077]: Remove User Preference Code. Will submit in a separate PR


* [feat]: [PL-31077]: Fix bug with remove Account mapping for an accont


* [feat]: [PL-31077]: Fix bug with remove Account mapping for an accont


* [feat]: [PL-31077]: Change USER_SOURCE to UserSource and minor test cases update.


* [feat]: [PL-31077]: Change USER_SOURCE to UserSource and minor test cases update.


* [feat]: [PL-31077]: Change USER_SOURCE to UserSource and minor test cases update.


* [feat]: [PL-31077]: Change USER_SOURCE to UserSource and minor test cases update.


* [feat]: [PL-31077]: Added Tests


* [feat]: [PL-31077]: Fixed Tests


* [feat]: [PL-31077]: Fixed Tests


* [feat]: [PL-31077]: Fixed removeUserAccountLevelDataForThisAccount


* [feat]: [PL-31077]: Fixed Tests


---------

+++ 400-rest/src/main/java/software/wings/resources/UserResourceNG.java
@@ -10,6 +10,7 @@ package software.wings.resources;
 import static io.harness.beans.PageResponse.PageResponseBuilder.aPageResponse;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
+import static io.harness.ng.core.common.beans.Generation.NG;
 import static io.harness.security.dto.PrincipalType.USER;
 
 import static software.wings.security.PermissionAttribute.PermissionType.USER_PERMISSION_MANAGEMENT;
@@ -111,7 +112,7 @@ public class UserResourceNG {
     User user = convertUserRequesttoUser(userRequest);
     String accountId = user.getDefaultAccountId();
 
-    User createdUser = userService.createNewUserAndSignIn(user, accountId);
+    User createdUser = userService.createNewUserAndSignIn(user, accountId, NG);
 
     return new RestResponse<>(convertUserToNgUser(createdUser));
   }
@@ -198,6 +199,10 @@ public class UserResourceNG {
   public RestResponse<Boolean> deleteUser(
       @QueryParam("accountId") String accountId, @QueryParam("userId") String userId) {
     if (featureFlagService.isEnabled(FeatureName.PL_USER_DELETION_V2, accountId)) {
+      if (featureFlagService.isEnabled(FeatureName.PL_USER_ACCOUNT_LEVEL_DATA_FLOW, accountId)
+          && userService.delete(accountId, userId, NG)) {
+        return new RestResponse<>(true);
+      }
       if (userService.isUserPresent(userId) && userServiceHelper.isUserActiveInNG(userService.get(userId), accountId)) {
         userServiceHelper.deleteUserFromNG(userId, accountId, NGRemoveUserFilter.ACCOUNT_LAST_ADMIN_CHECK);
       }
@@ -317,6 +322,14 @@ public class UserResourceNG {
       @QueryParam("isScimInvite") boolean isScimInvite,
       @QueryParam("shouldSendTwoFactorAuthResetEmail") boolean shouldSendTwoFactorAuthResetEmail) {
     userService.completeNGInvite(userInvite, isScimInvite, shouldSendTwoFactorAuthResetEmail);
+    return new RestResponse<>(true);
+  }
+
+  @PUT
+  @Path("invites/user")
+  public RestResponse<Boolean> createUserWithAccountLevelDataForInvite(@Body @NotNull UserInviteDTO userInvite,
+      @QueryParam("shouldSendTwoFactorAuthResetEmail") boolean shouldSendTwoFactorAuthResetEmail) {
+    userService.completeNGInviteWithAccountLevelData(userInvite, shouldSendTwoFactorAuthResetEmail);
     return new RestResponse<>(true);
   }
